Product and category link now in a combined file
